+++ source/Handlebars/Compiler/HandlebarsCompiler.cs
@@ -72,47 +72,6 @@ namespace HandlebarsDotNet.Compiler
                 compiledLayout(writer, bindingContext);
             };
         }
-
-        internal class DynamicViewModel : DynamicObject
-        {
-            private readonly object[] _objects;
-            private static readonly BindingFlags BindingFlags = BindingFlags.Public | BindingFlags.Instance | BindingFlags.IgnoreCase;
-
-            public DynamicViewModel(params object[] objects)
-            {
-                _objects = objects;
-            }
-
-            public override IEnumerable<string> GetDynamicMemberNames()
-            {
-                return _objects.Select(o => o.GetType())
-                    .SelectMany(t => t.GetMembers(BindingFlags))
-                    .Select(m => m.Name);
-            }
-
-            public override bool TryGetMember(GetMemberBinder binder, out object result)
-            {
-                result = null;
-                foreach (var target in _objects)
-                {
-                    var member = target.GetType().GetMember(binder.Name, BindingFlags);
-                    if (member.Length > 0)
-                    {
-                        if (member[0] is PropertyInfo)
-                        {
-                            result = ((PropertyInfo)member[0]).GetValue(target, null);
-                            return true;
-                        }
-                        if (member[0] is FieldInfo)
-                        {
-                            result = ((FieldInfo)member[0]).GetValue(target);
-                            return true;
-                        }
-                    }
-                }
-                return false;
-            }
-        }
     }
 }
